+++ docs/main.main.js
@@ -16,7 +16,7 @@
 /***/ ((module, __webpack_exports__, __webpack_require__) => {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var _node_modules_css_loader_dist_runtime_noSourceMaps_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../../node_modules/css-loader/dist/runtime/noSourceMaps.js */ \"./node_modules/css-loader/dist/runtime/noSourceMaps.js\");\n/* harmony import */ var _node_modules_css_loader_dist_runtime_noSourceMaps_js__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(_node_modules_css_loader_dist_runtime_noSourceMaps_js__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var _node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../node_modules/css-loader/dist/runtime/api.js */ \"./node_modules/css-loader/dist/runtime/api.js\");\n/* harmony import */ var _node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(_node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1__);\n// Imports\n\n\nvar ___CSS_LOADER_EXPORT___ = _node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1___default()((_node_modules_css_loader_dist_runtime_noSourceMaps_js__WEBPACK_IMPORTED_MODULE_0___default()));\n// Module\n___CSS_LOADER_EXPORT___.push([module.id, \".header-nav {\\n  position: sticky;\\n  height: 50px;\\n  display: flex;\\n  padding: 5px;\\n  justify-content: space-between;\\n  padding-bottom: 50px;\\n}\\n\\n.logo {\\n  border-radius: 50%;\\n  width: 50px;\\n  height: 50px;\\n  border: 4px solid rgb(73, 73, 73);\\n  display: flex;\\n  justify-content: center;\\n  align-items: center;\\n  color: rgb(214, 214, 214);\\n  background: rgb(165, 164, 164);\\n  box-shadow: 2px 4px 8px gray;\\n  font-size: 30px;\\n}\\n\\n.for-logo-img {\\n  width: 80px;\\n  padding: 12px;\\n}\\n\\n.nav-links {\\n  display: flex;\\n  gap: 30px;\\n  padding: 2px;\\n}\\n.nav-links a {\\n  color: rgb(214, 130, 130);\\n  font-size: 28px;\\n  font-weight: 800;\\n  display: flex;\\n}\\n\\n.meal-count {\\n  font-size: 18px;\\n  height: max-content;\\n  padding: 4px;\\n  margin-left: 4px;\\n  background: rgb(214, 130, 130);\\n  color: white;\\n  border-radius: 8px;\\n}\\n\\n.main-menu {\\n  position: fixed;\\n  right: -110%;\\n  gap: 0 !important;\\n  top: 90px;\\n  flex-direction: column;\\n  background-color: rgba(255, 255, 255, 0.8);\\n  width: 100%;\\n  text-align: left !important;\\n  transition: 0.3s;\\n  align-items: flex-end !important;\\n  height: 100vh;\\n  z-index: 999;\\n  list-style-type: none !important;\\n}\\n.main-menu.active {\\n  right: 0;\\n  margin-top: 0;\\n}\\n.main-menu .nav-item .nav-link {\\n  padding: 10px;\\n  text-decoration: none;\\n}\\n\\n.menu-ico {\\n  display: block;\\n}\\n\\n.bar {\\n  display: block;\\n  width: 24px;\\n  height: 3px;\\n  margin: 5px auto;\\n  transition: all 0.3s ease;\\n  background-color: rgb(214, 130, 130);\\n}\\n\\n.menu-ico.active .bar:nth-child(2) {\\n  opacity: 0;\\n}\\n.menu-ico.active .bar:nth-child(1) {\\n  transform: translateY(8px) rotate(45deg);\\n  background-color: rgb(214, 130, 130);\\n}\\n.menu-ico.active .bar:nth-child(3) {\\n  transform: translateY(-8px) rotate(-45deg);\\n  background-color: rgb(214, 130, 130);\\n}\\n\\n.danger {\\n  color: red;\\n}\\n\\n.test {\\n  border: 2px solid red;\\n  padding: 24px;\\n  display: flex;\\n  justify-content: space-around;\\n}\\n\\nheader {\\n  display: flex;\\n  align-items: center;\\n  justify-content: center;\\n}\\n\\nmain {\\n  width: 100%;\\n  min-height: 100vh;\\n  display: flex;\\n  flex-direction: column;\\n  justify-content: space-between;\\n  align-items: stretch;\\n}\\n\\nsection {\\n  display: flex;\\n  flex-direction: column;\\n  justify-content: center;\\n  align-items: center;\\n  padding: 10px;\\n}\\n\\n.container {\\n  width: 95%;\\n}\\n\\n.placeholder {\\n  width: 300px;\\n  height: 250px;\\n  background-color: grey;\\n  display: flex;\\n  justify-content: center;\\n  align-items: center;\\n  margin-bottom: 16px;\\n}\\n\\nbutton {\\n  padding: 8px;\\n}\\n\\n.user-name {\\n  font-weight: bold;\\n}\\n\\n.comment-date {\\n  font-size: small;\\n  font-style: italic;\\n}\\n\\n.comment-count {\\n  padding: 2px 4px;\\n  border-radius: 4px;\\n  background-color: black;\\n  color: white;\\n}\\n\\n.comment-form {\\n  display: flex;\\n  flex-direction: column;\\n  align-items: center;\\n}\\n.comment-form input,\\n.comment-form textarea {\\n  width: 80%;\\n  min-height: 36px;\\n  padding: 8px;\\n}\\n.comment-form input::placeholder,\\n.comment-form textarea::placeholder {\\n  font-size: 18px;\\n}\\n.comment-form .submit-btn {\\n  width: 50%;\\n  min-width: 150px;\\n  background-color: #167ec3;\\n  border: none;\\n  border-radius: 8px;\\n  padding: 8px;\\n  font-size: 18px;\\n  color: white;\\n  font-weight: bold;\\n}\\n\\nfooter {\\n  display: flex;\\n  flex-direction: column;\\n  align-items: center;\\n}\\nfooter .container {\\n  padding: 16px;\\n  border: none;\\n  margin-bottom: 24px;\\n  border-radius: 8px;\\n  background-color: #09568a;\\n  color: white;\\n  text-align: center;\\n}\\nfooter .container a {\\n  color: white;\\n  font-weight: bold;\\n}\\n\\n@media screen and (min-width: 768px) {\\n  .container {\\n    width: 80%;\\n  }\\n  .card-grid {\\n    display: grid;\\n    grid-template-columns: repeat(2, 1fr) !important;\\n    padding: 20px;\\n    column-gap: 24px;\\n  }\\n  .menu-ico {\\n    display: none;\\n  }\\n  .main-menu {\\n    position: relative !important;\\n    display: flex;\\n    flex-direction: row !important;\\n    width: 100%;\\n    text-align: right !important;\\n    transition: 0.3s;\\n    align-items: center !important;\\n    z-index: 999;\\n    list-style-type: none;\\n    justify-content: flex-end;\\n    right: 0 !important;\\n    background: transparent !important;\\n    top: 0 !important;\\n    height: auto !important;\\n  }\\n}\\n@media screen and (min-width: 960px) {\\n  .card-grid {\\n    display: grid;\\n    grid-template-columns: repeat(3, 1fr) !important;\\n  }\\n}\\n.card-grid {\\n  display: grid;\\n  grid-template-columns: 1fr;\\n  padding: 20px;\\n  column-gap: 24px;\\n}\\n\\n.card {\\n  display: flex;\\n  margin-bottom: 20px;\\n  background-color: #fff;\\n  flex-direction: column;\\n  justify-content: space-between;\\n  align-items: center;\\n  width: 100%;\\n  height: auto;\\n  border: 1px solid #ccc;\\n  border-radius: 4px;\\n  box-shadow: 2px 2px 8px rgba(0, 0, 0, 0.1);\\n  padding-bottom: 16px;\\n  overflow: hidden;\\n}\\n.card .image-placeholder {\\n  width: 100%;\\n  aspect-ratio: 4/3;\\n  background-color: #eee;\\n}\\n.card .text-flex {\\n  display: flex;\\n  flex-direction: column;\\n  align-items: center;\\n  padding: 16px;\\n}\\n.card .text-flex p {\\n  margin: 0;\\n  font-size: 18px;\\n  font-weight: bold;\\n  text-align: center;\\n}\\n.card .for-heart {\\n  display: flex;\\n  align-items: center;\\n  margin-top: 8px;\\n  color: #e74c3c;\\n}\\n.card .for-heart i {\\n  font-style: normal;\\n  margin-left: 4px;\\n  font-size: 14px;\\n}\\n.card .for-btn {\\n  display: flex;\\n  align-items: center;\\n  justify-content: center;\\n  margin-top: 8px;\\n}\\n.card .comment-btn {\\n  width: 100%;\\n  height: 32px;\\n  background-color: #3498db;\\n  border: none;\\n  border-radius: 4px;\\n  color: #fff;\\n  font-size: 14px;\\n  cursor: pointer;\\n}\\n.card .comment-btn:hover {\\n  background-color: #2980b9;\\n}\\n\\n.like-count {\\n  margin-left: 4px;\\n}\\n\\n.pop-up {\\n  position: fixed;\\n  top: 50%;\\n  left: 50%;\\n  transform: translate(-50%, -50%);\\n  width: 50%;\\n  height: 90%;\\n  background-color: #fff;\\n  border-radius: 4px;\\n  box-shadow: 2px 2px 8px rgba(0, 0, 0, 0.1);\\n  padding: 16px;\\n  overflow-y: scroll;\\n}\\n.pop-up .image-placeholder {\\n  width: 100%;\\n  aspect-ratio: 5/3;\\n  border: none;\\n  background-color: #eee;\\n  object-fit: cover;\\n  object-position: center;\\n}\\n.pop-up h2 {\\n  margin: 0;\\n  font-size: 24px;\\n  font-weight: bold;\\n  text-align: center;\\n}\\n.pop-up ul {\\n  list-style: none;\\n  margin: 0;\\n  padding: 0;\\n}\\n.pop-up ul li {\\n  margin: 8px 0;\\n  font-size: 18px;\\n}\\n\\n.pop-up-bg {\\n  width: 100vw;\\n  height: 100vh;\\n  position: fixed;\\n  top: 0;\\n  left: 0;\\n  background-color: rgba(212, 212, 212, 0.5);\\n}\\n\\n.exit-btn {\\n  position: absolute;\\n  display: flex;\\n  align-items: center;\\n  justify-content: center;\\n  text-align: center;\\n  top: 16px;\\n  right: 36px;\\n  width: 36px;\\n  height: 36px;\\n  background-color: transparent;\\n  border: none;\\n  cursor: pointer;\\n  font-size: 20px;\\n  font-weight: bold;\\n  color: grey;\\n  border-radius: 50%;\\n  transition: all 500ms ease;\\n}\\n.exit-btn:hover {\\n  color: white;\\n  background-color: rgb(30, 30, 30);\\n}\\n\\n.comments-heading {\\n  width: max-content;\\n  padding: 12px;\\n  background-color: rgb(30, 30, 30);\\n  color: white;\\n}\\n\\n.comments-count {\\n  color: rgb(30, 30, 30);\\n  background-color: white;\\n  line-height: 1;\\n  padding: 2px 8px;\\n  border-radius: 25%;\\n}\\n\\n* {\\n  text-decoration: none;\\n  font-family: \\\"poppins\\\", sans-serif;\\n}\", \"\"]);\n// Exports\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (___CSS_LOADER_EXPORT___);\n\n\n//# sourceURL=webpack://webpack-demo/./src/style/style.scss?./node_modules/css-loader/dist/cjs.js!./node_modules/sass-loader/dist/cjs.js");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var _node_modules_css_loader_dist_runtime_noSourceMaps_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../../node_modules/css-loader/dist/runtime/noSourceMaps.js */ \"./node_modules/css-loader/dist/runtime/noSourceMaps.js\");\n/* harmony import */ var _node_modules_css_loader_dist_runtime_noSourceMaps_js__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(_node_modules_css_loader_dist_runtime_noSourceMaps_js__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var _node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../node_modules/css-loader/dist/runtime/api.js */ \"./node_modules/css-loader/dist/runtime/api.js\");\n/* harmony import */ var _node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(_node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1__);\n// Imports\n\n\nvar ___CSS_LOADER_EXPORT___ = _node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1___default()((_node_modules_css_loader_dist_runtime_noSourceMaps_js__WEBPACK_IMPORTED_MODULE_0___default()));\n// Module\n___CSS_LOADER_EXPORT___.push([module.id, \".header-nav {\\n  position: sticky;\\n  height: 50px;\\n  display: flex;\\n  padding: 5px;\\n  justify-content: space-between;\\n  padding-bottom: 50px;\\n}\\n\\n.logo {\\n  border-radius: 50%;\\n  width: 50px;\\n  height: 50px;\\n  border: 4px solid rgb(73, 73, 73);\\n  display: flex;\\n  justify-content: center;\\n  align-items: center;\\n  color: rgb(214, 214, 214);\\n  background: rgb(165, 164, 164);\\n  box-shadow: 2px 4px 8px gray;\\n  font-size: 30px;\\n}\\n\\n.for-logo-img {\\n  width: 80px;\\n  padding: 12px;\\n}\\n\\n.nav-links {\\n  display: flex;\\n  gap: 30px;\\n  padding: 2px;\\n}\\n.nav-links a {\\n  color: rgb(214, 130, 130);\\n  font-size: 28px;\\n  font-weight: 800;\\n  display: flex;\\n}\\n\\n.meal-count {\\n  font-size: 18px;\\n  height: max-content;\\n  padding: 4px;\\n  margin-left: 4px;\\n  background: rgb(214, 130, 130);\\n  color: white;\\n  border-radius: 8px;\\n}\\n\\n.main-menu {\\n  position: fixed;\\n  right: -110%;\\n  gap: 0 !important;\\n  top: 90px;\\n  flex-direction: column;\\n  background-color: rgba(255, 255, 255, 0.8);\\n  width: 100%;\\n  text-align: left !important;\\n  transition: 0.3s;\\n  align-items: flex-end !important;\\n  height: 100vh;\\n  z-index: 999;\\n  list-style-type: none !important;\\n}\\n.main-menu.active {\\n  right: 0;\\n  margin-top: 0;\\n}\\n.main-menu .nav-item .nav-link {\\n  padding: 10px;\\n  text-decoration: none;\\n}\\n\\n.menu-ico {\\n  display: block;\\n}\\n\\n.bar {\\n  display: block;\\n  width: 24px;\\n  height: 3px;\\n  margin: 5px auto;\\n  transition: all 0.3s ease;\\n  background-color: rgb(214, 130, 130);\\n}\\n\\n.menu-ico.active .bar:nth-child(2) {\\n  opacity: 0;\\n}\\n.menu-ico.active .bar:nth-child(1) {\\n  transform: translateY(8px) rotate(45deg);\\n  background-color: rgb(214, 130, 130);\\n}\\n.menu-ico.active .bar:nth-child(3) {\\n  transform: translateY(-8px) rotate(-45deg);\\n  background-color: rgb(214, 130, 130);\\n}\\n\\n.danger {\\n  color: red;\\n}\\n\\n.test {\\n  border: 2px solid red;\\n  padding: 24px;\\n  display: flex;\\n  justify-content: space-around;\\n}\\n\\nheader {\\n  display: flex;\\n  align-items: center;\\n  justify-content: center;\\n}\\n\\nmain {\\n  width: 100%;\\n  min-height: 100vh;\\n  display: flex;\\n  flex-direction: column;\\n  justify-content: space-between;\\n  align-items: stretch;\\n}\\n\\nsection {\\n  display: flex;\\n  flex-direction: column;\\n  justify-content: center;\\n  align-items: center;\\n  padding: 10px;\\n}\\n\\n.container {\\n  width: 95%;\\n}\\n\\n.placeholder {\\n  width: 300px;\\n  height: 250px;\\n  background-color: grey;\\n  display: flex;\\n  justify-content: center;\\n  align-items: center;\\n  margin-bottom: 16px;\\n}\\n\\nbutton {\\n  padding: 8px;\\n}\\n\\n.user-name {\\n  font-weight: bold;\\n}\\n\\n.comment-date {\\n  font-size: small;\\n  font-style: italic;\\n}\\n\\n.comment-count {\\n  padding: 2px 4px;\\n  border-radius: 4px;\\n  background-color: black;\\n  color: white;\\n}\\n\\n.comment-form {\\n  display: flex;\\n  flex-direction: column;\\n  align-items: center;\\n}\\n.comment-form input,\\n.comment-form textarea {\\n  width: 80%;\\n  min-height: 36px;\\n  padding: 8px;\\n}\\n.comment-form input::placeholder,\\n.comment-form textarea::placeholder {\\n  font-size: 18px;\\n}\\n.comment-form .submit-btn {\\n  width: 50%;\\n  min-width: 150px;\\n  background-color: #167ec3;\\n  border: none;\\n  border-radius: 8px;\\n  padding: 8px;\\n  font-size: 18px;\\n  color: white;\\n  font-weight: bold;\\n}\\n\\nfooter {\\n  display: flex;\\n  flex-direction: column;\\n  align-items: center;\\n}\\nfooter .container {\\n  padding: 16px;\\n  border: none;\\n  margin-bottom: 24px;\\n  border-radius: 8px;\\n  background-color: #09568a;\\n  color: white;\\n  text-align: center;\\n}\\nfooter .container a {\\n  color: white;\\n  font-weight: bold;\\n}\\n\\n@media screen and (min-width: 768px) {\\n  .container {\\n    width: 80%;\\n  }\\n  .card-grid {\\n    display: grid;\\n    grid-template-columns: repeat(2, 1fr) !important;\\n    padding: 20px;\\n    column-gap: 24px;\\n  }\\n  .menu-ico {\\n    display: none;\\n  }\\n  .main-menu {\\n    position: relative !important;\\n    display: flex;\\n    flex-direction: row !important;\\n    width: 100%;\\n    text-align: right !important;\\n    transition: 0.3s;\\n    align-items: center !important;\\n    z-index: 999;\\n    list-style-type: none;\\n    justify-content: flex-end;\\n    right: 0 !important;\\n    background: transparent !important;\\n    top: 0 !important;\\n    height: auto !important;\\n  }\\n  .pop-up {\\n    width: 50% !important;\\n    height: 90% !important;\\n  }\\n}\\n@media screen and (min-width: 960px) {\\n  .card-grid {\\n    display: grid;\\n    grid-template-columns: repeat(3, 1fr) !important;\\n  }\\n}\\n.card-grid {\\n  display: grid;\\n  grid-template-columns: 1fr;\\n  padding: 20px;\\n  column-gap: 24px;\\n}\\n\\n.card {\\n  display: flex;\\n  margin-bottom: 20px;\\n  background-color: #fff;\\n  flex-direction: column;\\n  justify-content: space-between;\\n  align-items: center;\\n  width: 100%;\\n  height: auto;\\n  border: 1px solid #ccc;\\n  border-radius: 4px;\\n  box-shadow: 2px 2px 8px rgba(0, 0, 0, 0.1);\\n  padding-bottom: 16px;\\n  overflow: hidden;\\n}\\n.card .image-placeholder {\\n  width: 100%;\\n  aspect-ratio: 4/3;\\n  background-color: #eee;\\n}\\n.card .text-flex {\\n  display: flex;\\n  flex-direction: column;\\n  align-items: center;\\n  padding: 16px;\\n}\\n.card .text-flex p {\\n  margin: 0;\\n  font-size: 18px;\\n  font-weight: bold;\\n  text-align: center;\\n}\\n.card .for-heart {\\n  display: flex;\\n  align-items: center;\\n  margin-top: 8px;\\n  color: #e74c3c;\\n}\\n.card .for-heart i {\\n  font-style: normal;\\n  margin-left: 4px;\\n  font-size: 14px;\\n}\\n.card .for-btn {\\n  display: flex;\\n  align-items: center;\\n  justify-content: center;\\n  margin-top: 8px;\\n}\\n.card .comment-btn {\\n  width: 100%;\\n  height: 32px;\\n  background-color: #3498db;\\n  border: none;\\n  border-radius: 4px;\\n  color: #fff;\\n  font-size: 14px;\\n  cursor: pointer;\\n}\\n.card .comment-btn:hover {\\n  background-color: #2980b9;\\n}\\n\\n.like-count {\\n  margin-left: 4px;\\n}\\n\\n.pop-up {\\n  position: fixed;\\n  top: 50%;\\n  left: 50%;\\n  transform: translate(-50%, -50%);\\n  width: 70%;\\n  height: 70%;\\n  background-color: #fff;\\n  border-radius: 4px;\\n  box-shadow: 2px 2px 8px rgba(0, 0, 0, 0.1);\\n  padding: 16px;\\n  overflow-y: scroll;\\n}\\n.pop-up .image-placeholder {\\n  width: 100%;\\n  aspect-ratio: 5/3;\\n  border: none;\\n  background-color: #eee;\\n  object-fit: cover;\\n  object-position: center;\\n}\\n.pop-up h2 {\\n  margin: 0;\\n  font-size: 24px;\\n  font-weight: bold;\\n  text-align: center;\\n}\\n.pop-up ul {\\n  list-style: none;\\n  margin: 0;\\n  padding: 0;\\n}\\n.pop-up ul li {\\n  margin: 8px 0;\\n  font-size: 18px;\\n}\\n\\n.pop-up-bg {\\n  width: 100vw;\\n  height: 100vh;\\n  position: fixed;\\n  top: 0;\\n  left: 0;\\n  background-color: rgba(212, 212, 212, 0.5);\\n}\\n\\n.exit-btn {\\n  position: absolute;\\n  display: flex;\\n  align-items: center;\\n  justify-content: center;\\n  text-align: center;\\n  top: 16px;\\n  right: 36px;\\n  width: 36px;\\n  height: 36px;\\n  background-color: transparent;\\n  border: none;\\n  cursor: pointer;\\n  font-size: 20px;\\n  font-weight: bold;\\n  color: grey;\\n  border-radius: 50%;\\n  transition: all 500ms ease;\\n}\\n.exit-btn:hover {\\n  color: white;\\n  background-color: rgb(30, 30, 30);\\n}\\n\\n.comments-heading {\\n  width: max-content;\\n  padding: 12px;\\n  background-color: rgb(30, 30, 30);\\n  color: white;\\n}\\n\\n.comments-count {\\n  color: rgb(30, 30, 30);\\n  background-color: white;\\n  line-height: 1;\\n  padding: 2px 8px;\\n  border-radius: 25%;\\n}\\n\\n* {\\n  text-decoration: none;\\n  font-family: \\\"poppins\\\", sans-serif;\\n}\", \"\"]);\n// Exports\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (___CSS_LOADER_EXPORT___);\n\n\n//# sourceURL=webpack://webpack-demo/./src/style/style.scss?./node_modules/css-loader/dist/cjs.js!./node_modules/sass-loader/dist/cjs.js");
 
 /***/ }),
 
@@ -151,6 +151,17 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 
 /***/ }),
 
+/***/ "./src/modules/commentcount.js":
+/*!*************************************!*\
+  !*** ./src/modules/commentcount.js ***!
+  \*************************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\nconst commentCount = (arr) => {\n  let result = 0;\n  if (arr.length === undefined) {\n    result = 0;\n  } else {\n    result = arr.length;\n  }\n  return result;\n};\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (commentCount);\n\n\n//# sourceURL=webpack://webpack-demo/./src/modules/commentcount.js?");
+
+/***/ }),
+
 /***/ "./src/modules/comments.js":
 /*!*********************************!*\
   !*** ./src/modules/comments.js ***!
@@ -158,7 +169,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"addComment\": () => (/* binding */ addComment),\n/* harmony export */   \"commentCount\": () => (/* binding */ commentCount),\n/* harmony export */   \"renderComment\": () => (/* binding */ renderComment),\n/* harmony export */   \"showComment\": () => (/* binding */ showComment),\n/* harmony export */   \"showCount\": () => (/* binding */ showCount)\n/* harmony export */ });\nconst addComment = async (baseLink, itemId) => {\n  const user = document.getElementById(`username-${itemId}`);\n  const text = document.getElementById(`user-comment-${itemId}`);\n  await fetch(`${baseLink}/comments/`, {\n    method: 'POST',\n    headers: {\n      'Content-type': 'application/json; charset=UTF-8',\n    },\n    body: JSON.stringify({\n      item_id: itemId,\n      username: user.value,\n      comment: text.value,\n    }),\n  });\n  user.value = '';\n  text.value = '';\n};\nconst renderComment = (arr) => {\n  let comments = '';\n  if (arr.length !== undefined) {\n    arr.forEach((obj) => {\n      comments += `<p class=\"single-comment\"><span class=\"comment-date\">${obj.creation_date}</span><br><span class=\"user-name\">${obj.username}:</span> <span class=\"comment-text\">${obj.comment}</span></p>`;\n    });\n  }\n\n  return comments;\n};\nconst showComment = async (baseLink, itemId) => {\n  const response = await fetch(`${baseLink}/comments?item_id=${itemId}`);\n  const commentShow = document.querySelector(`#pop-${itemId} .show-comments`);\n  await response.json().then((comments) => {\n    commentShow.replaceChildren();\n    commentShow.innerHTML = renderComment(comments);\n  });\n};\nconst commentCount = (arr) => {\n  let result = 0;\n  if (arr.length === undefined) {\n    result = 0;\n  } else {\n    result = arr.length;\n  }\n  return result;\n};\nconst showCount = async (baseLink, itemId) => {\n  const response = await fetch(`${baseLink}/comments?item_id=${itemId}`);\n  const countSpan = document.getElementById(`#comment-count-${itemId}`);\n  await response.json().then((comments) => {\n    countSpan.replaceChildren();\n    countSpan.innerHTML = commentCount(comments);\n  });\n};\n\n\n\n//# sourceURL=webpack://webpack-demo/./src/modules/comments.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"addComment\": () => (/* binding */ addComment),\n/* harmony export */   \"renderComment\": () => (/* binding */ renderComment),\n/* harmony export */   \"showComment\": () => (/* binding */ showComment),\n/* harmony export */   \"showCount\": () => (/* binding */ showCount)\n/* harmony export */ });\n/* harmony import */ var _commentcount_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./commentcount.js */ \"./src/modules/commentcount.js\");\n\nconst addComment = async (baseLink, itemId) => {\n  const user = document.getElementById(`username-${itemId}`);\n  const text = document.getElementById(`user-comment-${itemId}`);\n  await fetch(`${baseLink}/comments/`, {\n    method: \"POST\",\n    headers: {\n      \"Content-type\": \"application/json; charset=UTF-8\",\n    },\n    body: JSON.stringify({\n      item_id: itemId,\n      username: user.value,\n      comment: text.value,\n    }),\n  });\n  user.value = \"\";\n  text.value = \"\";\n};\nconst renderComment = (arr) => {\n  let comments = \"\";\n  if (arr.length !== undefined) {\n    arr.forEach((obj) => {\n      comments += `<p class=\"single-comment\"><span class=\"comment-date\">${obj.creation_date}</span><br><span class=\"user-name\">${obj.username}:</span> <span class=\"comment-text\">${obj.comment}</span></p>`;\n    });\n  }\n\n  return comments;\n};\nconst showComment = async (baseLink, itemId) => {\n  const response = await fetch(`${baseLink}/comments?item_id=${itemId}`);\n  const commentShow = document.querySelector(`#pop-${itemId} .show-comments`);\n  await response.json().then((comments) => {\n    commentShow.replaceChildren();\n    commentShow.innerHTML = renderComment(comments);\n  });\n};\n\nconst showCount = async (baseLink, itemId) => {\n  const response = await fetch(`${baseLink}/comments?item_id=${itemId}`);\n  const countSpan = document.getElementById(`#comment-count-${itemId}`);\n  await response.json().then((comments) => {\n    countSpan.replaceChildren();\n    countSpan.innerHTML = (0,_commentcount_js__WEBPACK_IMPORTED_MODULE_0__[\"default\"])(comments);\n  });\n};\n\n\n\n//# sourceURL=webpack://webpack-demo/./src/modules/comments.js?");
 
 /***/ }),
 
@@ -170,6 +181,17 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 
 "use strict";
 eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"appId\": () => (/* binding */ appId),\n/* harmony export */   \"baseLink\": () => (/* binding */ baseLink),\n/* harmony export */   \"catBeef\": () => (/* binding */ catBeef),\n/* harmony export */   \"catDessert\": () => (/* binding */ catDessert),\n/* harmony export */   \"catSeafood\": () => (/* binding */ catSeafood),\n/* harmony export */   \"menuIco\": () => (/* binding */ menuIco),\n/* harmony export */   \"navMenu\": () => (/* binding */ navMenu)\n/* harmony export */ });\nconst appId = 'No6xjeOV6L9eg8TkvJgU';\nconst baseLink = `https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/${appId}`;\nconst menuIco = document.querySelector('.menu-ico');\nconst navMenu = document.querySelector('.main-menu');\nconst catBeef = document.getElementById('cat-beef');\nconst catSeafood = document.getElementById('cat-seafood');\nconst catDessert = document.getElementById('cat-dessert');\n\n\n\n\n//# sourceURL=webpack://webpack-demo/./src/modules/globals.js?");
+
+/***/ }),
+
+/***/ "./src/modules/itemcount.js":
+/*!**********************************!*\
+  !*** ./src/modules/itemcount.js ***!
+  \**********************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\nconst mealsCount = (arr) => {\n  let result = 0;\n  if (arr.length === undefined) {\n    result = 0;\n  } else {\n    result = arr.length;\n  }\n  return result;\n};\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (mealsCount);\n\n\n//# sourceURL=webpack://webpack-demo/./src/modules/itemcount.js?");
 
 /***/ }),
 
@@ -191,7 +213,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"countMeals\": () => (/* binding */ countMeals),\n/* harmony export */   \"mealsCount\": () => (/* binding */ mealsCount),\n/* harmony export */   \"showMeals\": () => (/* binding */ showMeals)\n/* harmony export */ });\n/* harmony import */ var _card_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./card.js */ \"./src/modules/card.js\");\n/* harmony import */ var _globals_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./globals.js */ \"./src/modules/globals.js\");\n/* harmony import */ var _like_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./like.js */ \"./src/modules/like.js\");\n\n\n\n\nconst showMeals = async (category) => {\n  const response = await fetch(\n    `https://www.themealdb.com/api/json/v1/1/filter.php?c=${category}`,\n  );\n  await response.json().then((food) => {\n    const foods = food.meals;\n    (0,_card_js__WEBPACK_IMPORTED_MODULE_0__[\"default\"])(foods);\n  });\n  (0,_like_js__WEBPACK_IMPORTED_MODULE_2__.likeCount)(_globals_js__WEBPACK_IMPORTED_MODULE_1__.baseLink);\n};\nconst mealsCount = (arr) => {\n  let result = 0;\n  if (arr.length === undefined) {\n    result = 0;\n  } else {\n    result = arr.length;\n  }\n  return result;\n};\nconst countMeals = async (category, menuItem) => {\n  const response = await fetch(\n    `https://www.themealdb.com/api/json/v1/1/filter.php?c=${category}`,\n  );\n  await response.json().then((food) => {\n    const foods = food.meals;\n    menuItem.childNodes[1].innerText = mealsCount(foods);\n  });\n};\n\n\n\n//# sourceURL=webpack://webpack-demo/./src/modules/meals.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"countMeals\": () => (/* binding */ countMeals),\n/* harmony export */   \"showMeals\": () => (/* binding */ showMeals)\n/* harmony export */ });\n/* harmony import */ var _card_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./card.js */ \"./src/modules/card.js\");\n/* harmony import */ var _globals_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./globals.js */ \"./src/modules/globals.js\");\n/* harmony import */ var _like_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./like.js */ \"./src/modules/like.js\");\n/* harmony import */ var _itemcount_js__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./itemcount.js */ \"./src/modules/itemcount.js\");\n\n\n\n\nconst showMeals = async (category) => {\n  const response = await fetch(\n    `https://www.themealdb.com/api/json/v1/1/filter.php?c=${category}`\n  );\n  await response.json().then((food) => {\n    const foods = food.meals;\n    (0,_card_js__WEBPACK_IMPORTED_MODULE_0__[\"default\"])(foods);\n  });\n  (0,_like_js__WEBPACK_IMPORTED_MODULE_2__.likeCount)(_globals_js__WEBPACK_IMPORTED_MODULE_1__.baseLink);\n};\n\nconst countMeals = async (category, menuItem) => {\n  const response = await fetch(\n    `https://www.themealdb.com/api/json/v1/1/filter.php?c=${category}`\n  );\n  await response.json().then((food) => {\n    const foods = food.meals;\n    menuItem.childNodes[1].innerText = (0,_itemcount_js__WEBPACK_IMPORTED_MODULE_3__[\"default\"])(foods);\n  });\n};\n\n\n\n//# sourceURL=webpack://webpack-demo/./src/modules/meals.js?");
 
 /***/ }),
