+++ src/modules/itemcount.js
@@ -1,0 +1,10 @@
+const mealsCount = (arr) => {
+  let result = 0;
+  if (arr.length === undefined) {
+    result = 0;
+  } else {
+    result = arr.length;
+  }
+  return result;
+};
+export default mealsCount;
