+++ src/modules/meals.js
@@ -1,6 +1,7 @@
 import renderCards from './card.js';
 import { baseLink } from './globals.js';
 import { likeCount } from './like.js';
+import mealsCount from './itemcount.js';
 
 const showMeals = async (category) => {
   const response = await fetch(
@@ -12,15 +13,7 @@ const showMeals = async (category) => {
   });
   likeCount(baseLink);
 };
-const mealsCount = (arr) => {
-  let result = 0;
-  if (arr.length === undefined) {
-    result = 0;
-  } else {
-    result = arr.length;
-  }
-  return result;
-};
+
 const countMeals = async (category, menuItem) => {
   const response = await fetch(
     `https://www.themealdb.com/api/json/v1/1/filter.php?c=${category}`,
@@ -30,4 +23,4 @@ const countMeals = async (category, menuItem) => {
     menuItem.childNodes[1].innerText = mealsCount(foods);
   });
 };
-export { showMeals, countMeals, mealsCount };
+export { showMeals, countMeals };
