+++ src/style/_breakpoints.scss
@@ -30,6 +30,11 @@
     top: 0 !important;
     height: auto !important;
   }
+
+  .pop-up {
+    width: 50% !important;
+    height: 90% !important;
+  }
 }
 
 @media screen and (min-width: 960px) {
